# Conflicts:
#	package-lock.json
#	package.json
#	yarn.lock

+++ spectron/util/util.ts
@@ -7,6 +7,7 @@ import {LocalRepository} from "../../electron/src/storage/types/local-repository
 import {UserRepository} from "../../electron/src/storage/types/user-repository";
 import rimraf = require("rimraf");
 import ITestCallbackContext = Mocha.ITestCallbackContext;
+import {encodeBase64} from "../../electron/src/security/encoder";
 
 interface FnTestConfig {
     localRepository: Partial<LocalRepository>,
@@ -17,9 +18,7 @@ interface FnTestConfig {
     }[],
     waitForMainWindow: boolean,
     testTimeout: number,
-    retries: number,
-    swapFiles: { userID: string, content: string }[]
-
+    retries: number
 }
 
 function isDevServer() {
@@ -31,7 +30,7 @@ function findAppBinary() {
     if (isDevServer()) {
         return glob.sync("./node_modules/.bin/electron")[0];
     } else if (process.platform.startsWith("win")) {
-        return path.resolve("./build/win-unpacked/rabix-composer.exe");
+        return glob.sync("./build/**/rabix-composer.exe")[0];
     } else if (process.platform.startsWith("darwin")) {
         return path.resolve("./build/mac/rabix-composer.app/Contents/MacOS/rabix-composer");
     } else {
@@ -42,7 +41,7 @@ function findAppBinary() {
 export function boot(context: ITestCallbackContext, testConfig: Partial<FnTestConfig> = {}): Promise<spectron.Application> {
 
     context.retries(testConfig.retries || 0);
-    context.timeout(testConfig.testTimeout || 5000);
+    context.timeout(testConfig.testTimeout || 30000);
 
     const testTitle      = context.test.fullTitle();
     const globalTestDir  = path.resolve(`${__dirname}/../../.testrun`);
@@ -53,19 +52,14 @@ export function boot(context: ITestCallbackContext, testConfig: Partial<FnTestCo
 
     testConfig.localRepository = Object.assign(new LocalRepository(), testConfig.localRepository || {});
 
-    fs.outputJSONSync(localProfilePath, testConfig.localRepository, {
-        spaces: 4,
-        replacer: null
-    });
+    fs.outputFileSync(localProfilePath, encodeBase64(JSON.stringify(testConfig.localRepository)));
 
     if (testConfig.platformRepositories) {
         for (let userID in testConfig.platformRepositories) {
             const profilePath = profilesDirPath + `/${userID}.json`;
             const profileData = Object.assign(new UserRepository(), testConfig.platformRepositories[userID] || {});
-            fs.outputJSONSync(profilePath, profileData, {
-                spaces: 4,
-                replacer: null
-            })
+
+            fs.outputFileSync(profilePath, encodeBase64(JSON.stringify(profileData)));
         }
     }
 
@@ -104,6 +98,9 @@ export function boot(context: ITestCallbackContext, testConfig: Partial<FnTestCo
 
 export function shutdown(app: spectron.Application) {
 
+    if(!app){
+        return;
+    }
 
     if (app.hasOwnProperty("testdir")) {
         rimraf.sync(app["testdir"]);
